clean up code - add commands to gen_test_data.sh - add compile_queue.py to library - factor out common code and apply linting
