+++ frontends/queues/tests/compiler/complex_tree_compile.py
@@ -1,5 +1,6 @@
-from compile_queue import build
+from queues.compile_queue import build
+
 """A test for the compiler-generated version of the same complex tree as complex_tree_test.py"""
 
 if __name__ == "__main__":
-  build("complex_tree.json").emit()
+    build("complex_tree.json").emit()
