+++ frontends/queues/tests/compiler/pifo_tree_compile.py
@@ -1,4 +1,5 @@
-from compile_queue import build
+from queues.compile_queue import build
+
 """
 A test for the compiler-generated version of the same pifo tree as in 
 pifo_tree_test.py (and the rio program rr_hier.sched in the packet
@@ -6,4 +7,4 @@ scheduling repo)
 """
 
 if __name__ == "__main__":
-  build("pifo_tree.json").emit()
+    build("pifo_tree.json").emit()
