+++ frontends/queues/tests/compiler/rr_compile.py
@@ -1,4 +1,5 @@
-from compile_queue import build
+from queues.compile_queue import build
+
 """
 A test for the compiler-generated version of the same round robin tree with 2 
 flows as in rr_2flows.py (and the rio program rr_2_classes.sched in the packet
@@ -6,4 +7,4 @@ scheduling repo)
 """
 
 if __name__ == "__main__":
-  build("rr_2flows.json").emit()
+    build("rr_2flows.json").emit()
